# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/dat-truoc": {
      "post": {
        "description": "Qua `sp_dat_truoc`. Bạn đọc đặt cho chính mình (bỏ `maNguoiDung`, nếu gửi mã người khác thì 403). Cán bộ đặt hộ thì phải gửi `maNguoiDung` (thiếu: 400). 422 khi: còn bản sách sẵn sàng (không cần đặt), đang mượn chính đầu sách đó, hoặc không đủ điều kiện đặt (người dùng không hoạt động, đang giữ sách quá hạn, còn nợ phạt, đã đặt rồi...).\n\n**Quyền:** mọi tài khoản đã đăng nhập.",
        "operationId": "DatTruoc_datTruoc",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatTruocDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatTruocDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đặt trước đầu sách",
        "tags": [
          "dat-truoc"
        ]
      },
      "get": {
        "description": "Bạn đọc luôn chỉ thấy lượt của mình (bỏ qua `maNguoiDung`). Cán bộ thấy tất cả và lọc được theo `maNguoiDung`.\n\n**Quyền:** mọi tài khoản đã đăng nhập.",
        "operationId": "DatTruoc_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Trang, bắt đầu từ 1.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số dòng mỗi trang (1–100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "trangThai",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CHO_XU_LY",
                "SAN_SANG_NHAN",
                "DA_NHAN",
                "HUY",
                "HET_HAN"
              ]
            }
          },
          {
            "name": "maNguoiDung",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DatTruocDto"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Danh sách lượt đặt trước",
        "tags": [
          "dat-truoc"
        ]
      }
    },
    "/dat-truoc/{maSach}": {
      "delete": {
        "description": "Qua `sp_huy_dat_truoc`. Bạn đọc hủy lượt của mình; cán bộ hủy hộ thì phải gửi query `maNguoiDung`.\n\n**Quyền:** mọi tài khoản đã đăng nhập.",
        "operationId": "DatTruoc_huy",
        "parameters": [
          {
            "name": "maSach",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maNguoiDung",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HuyDatTruocKetQuaDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hủy đặt trước",
        "tags": [
          "dat-truoc"
        ]
      }
    },
...
  },
  "info": {
    "title": "QLTV Nhóm 8 - API quản lý thư viện",
    "description": "API quản lý thư viện đại học (đồ án IE103, nhóm 8). Nghiệp vụ nằm trong MySQL (procedure, trigger, view); BE chỉ gọi lại và xác thực.\n\n**Cách dùng nhanh**\n1. `POST /auth/login` với tài khoản seed (ví dụ `ad001` / `AD001@Nhom8`), lấy `accessToken`.\n2. Bấm **Authorize** ở góc phải và dán token (không cần gõ chữ `Bearer`). Token được nhớ khi tải lại trang.\n3. Mỗi endpoint ghi rõ **Quyền** được gọi. Sai vai trò: 403. Token sai, hết hạn hoặc tài khoản bị khóa: 401.\n\n**Quy ước dữ liệu**\n- Khóa chính BIGINT (`id`) là **chuỗi**, ví dụ `\"13\"`. Các mã (`maSach`, `maBanSach`, `maNguoiDung`...) là chuỗi và là thứ dùng khi gọi API nghiệp vụ.\n- Endpoint danh sách có phân trang trả `{ data, total, page, limit }` (`page` từ 1, `limit` tối đa 100).\n- Cột ngày kiểu DATE trả dạng ISO `YYYY-MM-DDT00:00:00.000Z`.\n- Tiền (VND) từ báo cáo và `/me` là number; `soTien`/`giaBia` đọc từ bảng là chuỗi thập phân (Prisma Decimal).\n- Báo cáo (`/bao-cao/*`), `GET /sach` và `/me/*` giữ tên cột snake_case của view; các endpoint còn lại dùng camelCase.\n- Lỗi luôn có dạng `{ statusCode, message }`. 422 là lỗi nghiệp vụ do CSDL báo, `message` là thông báo cho người dùng (tiếng Việt không dấu).",
    "version": "2.2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 422
          },
          "error": {
            "type": "string",
            "example": "Unprocessable Entity",
            "description": "Tên mã HTTP (chỉ có ở lỗi do bộ lọc CSDL và một số lỗi của Nest)"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Nguoi dung dang no phat",
            "description": "Chuỗi thông báo; lỗi validate dữ liệu gửi lên (400) là mảng các thông báo"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "NguoiDungTomTatDto": {
        "type": "object",
        "properties": {
          "maNguoiDung": {
            "type": "string",
            "example": "SV001"
          },
          "hoTen": {
            "type": "string",
            "example": "Võ Hoàng Nhiên"
          }
        },
        "required": [
          "maNguoiDung",
          "hoTen"
        ]
      },
...
      "DatTruocDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "nguoiDungId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "sachId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "banSachId": {
            "type": "string",
            "nullable": true,
            "example": "1",
            "description": "Bản sách đang giữ cho lượt đặt (nếu có)"
          },
          "ngayDat": {
            "type": "string",
            "format": "date-time"
          },
          "hanGiu": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "Hạn giữ sách khi SAN_SANG_NHAN"
          },
          "trangThai": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/TrangThaiDatTruoc"
              }
            ]
          },
          "khoaDangHoatDong": {
            "type": "string",
            "nullable": true,
            "description": "Cột sinh tự động của DB (chống đặt trùng); FE bỏ qua"
          },
          "banSachDangGiu": {
            "type": "string",
            "nullable": true,
            "example": "1",
            "description": "Cột sinh tự động của DB; FE bỏ qua"
          },
          "sach": {
            "$ref": "#/components/schemas/DatTruocSachDto"
          },
          "nguoiDung": {
            "$ref": "#/components/schemas/NguoiDungTomTatDto"
          }
        },
        "required": [
          "id",
          "nguoiDungId",
          "sachId",
          "banSachId",
          "ngayDat",
          "hanGiu",
          "trangThai",
          "khoaDangHoatDong",
          "banSachDangGiu",
          "sach",
          "nguoiDung"
        ]
      },
      "TrangThaiDatTruoc": {
        "type": "string",
        "enum": [
          "CHO_XU_LY",
          "SAN_SANG_NHAN",
          "DA_NHAN",
          "HUY",
          "HET_HAN"
        ]
      },
      "DatTruocSachDto": {
        "type": "object",
        "properties": {
          "maSach": {
            "type": "string",
            "example": "S004"
          },
          "tenSach": {
            "type": "string"
          }
        },
        "required": [
          "maSach",
          "tenSach"
        ]
      },
      "HuyDatTruocKetQuaDto": {
        "type": "object",
        "properties": {
          "maSach": {
            "type": "string",
            "example": "S004"
          },
          "maNguoiDung": {
            "type": "string",
            "example": "SV002"
          },
          "trangThai": {
            "type": "string",
            "example": "HUY"
          }
        },
        "required": [
          "maSach",
          "maNguoiDung",
          "trangThai"
        ]
      },
...
    }
  }
}